Define PseudoRealRegisters using new RegDep enum
Certain registers defined inside RealRegisterEnum.hpp are
not hardware backed registers
(ex. TR::RealRegister::NumRegisters), but instead are used as
hints for the code generator or as dependencies for real
registers. This is quite prevalent in the z/codegen.This commit
separates these pseudo registers from the real register enum,
and updates the APIs where  needed. TR::RealRegister::MISC is
also deprecated because it represents the same thing as
TR::RealRegister::LastHPR.

Closes: #2491
Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>