Deprecate trj9 compiler directory
The runtime/compiler/trj9 directory currently serves no purpose as it's
parent directory is completely empty. In an effort to simplify the
codebase and closely match the directory layout of OMR we fold this
trj9 directory away so that understanding how OMR and OpenJ9 fit
together with respect to the compiler component is much easier.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>