Change custom-allocated maps to pointers
The `compiler/ilgen's` `MethodBuilder` and `TypeDictionary` classes
contains a number of class members, maps, that use a custom allocator
(`TR::typed_allocator<...> _memoryRegion`). The allocator itself are
deleted in the destructors of `MethodBuilder` and `TypeDictionary`:

_memoryRegion->~Region();
::operator delete(_memoryRegion, TR::Compiler->persistentAllocator());

Then, the class destructors automagically invokes the destructors for
member objects: `_structsByName`, `_unionsByName`, `_symbols`, etc.
Eventually, the method `_Node::_Freenode0` of the `_Tree_node` STL
class (std::map is implemented as a tree) for the head node will be
invoked. The method uses the allocator (`_memoryRegion` in our case, the
destroyed yet) to deallocate the head node and, since the allocator has
been destroyed yet in the destructor of the `MethodBuilder` or
`TypeDictionary` class, an access violation error appears:

[----------] 2 tests from JITILBuilderTest
[ RUN      ] JITILBuilderTest.ControlFlowTest
unknown file: error: SEH exception with code 0xc0000005 thrown in the
test body.
[  FAILED  ] JITILBuilderTest.ControlFlowTest (351 ms)

To avoid the situation, the `map` members of the `MethodBuilder` and
`TypeDictionary` classes are replaced with the pointers to maps and
the allocated memory is explicitly reclaimed in the destructors before
the allocator is being deleted.

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>