Enable aliasing of auto variables that have their address taken
In languages like C it is possible for local (auto) variables to have their address taken
and then passed by reference to another function. Example:

int foo(int*i, long long*l)
{
	*i = (int) *l;
	return 0;
}

int simplelocals(void)
{
	int a = 5;
	long long b = 42;
	foo(&a, &b);	// Just to fool the optimiser
	return a+(int)b;
}

Currently OMR's localCSE ignores the fact that the two locals were passed by reference to foo() and therefore
could have been changed by foo(). This results in incorrect optimization. The solution is to alias the autos
for "a" and "b" with the call "foo".

To enable this a new BitVector _addressTakenAutos is introduced in OMR::AliasBuilder. This is made available
to front-ends via OMR::AliasBuilder::addressTakenAutos(). The front-end is expected to set the symbols for
which addresses have been taken into this set.

The second part of this change is to use the aliasing information. The aliasing api that controls what the
rest of the optimizer sees is getUseDefAliasesBV in
https://github.com/eclipse/omr/blob/master/compiler/il/Aliases.cpp

In this routine the case for:

case TR::Symbol::IsResolvedMethod

Is updated to OR the addressTakenAliases with the method's aliases.

This change relates to the bug logged in https://github.com/eclipse/omr/issues/2768.

Signed-off-by: Dibyendu Majumdar <mobile@majumdar.org.uk>